requirements.txt
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
drf-writable-nested==0.6.3